+++ src/collective/documentgenerator/browser/generation_view.py
@@ -2,6 +2,7 @@
 
 from .. import _
 from AccessControl import Unauthorized
+from appy.pod.evaluator import Compromiser
 from appy.pod.renderer import CsvOptions
 from appy.pod.renderer import Renderer
 from appy.pod.styles_manager import TableProperties
@@ -127,13 +128,21 @@ class DocumentGenerationView(BrowserView):
         filename = self._get_filename()
         return rendered, filename, gen_context
 
-    def _get_filename(self):
-        """ """
-        # we limit filename to 120 characters
-        first_part = u'{0} {1}'.format(self.pod_template.title, safe_unicode(self.context.Title()))
+    def _get_filename(self, prepend_pod_title=True):
+        """
+        Return generated file filename.
+
+        :param prepend_pod_title: will preprend the POD template title before
+        context title for the generated file filename
+        title then pod template title
+        """
+        first_part = safe_unicode(self.context.Title())
+        if prepend_pod_title:
+            first_part = u'{0} {1}'.format(self.pod_template.title, first_part)
         # replace unicode special characters with ascii equivalent value
         first_part = unicodedata.normalize('NFKD', first_part).encode('ascii', 'ignore')
         util = queryUtility(IFileNameNormalizer)
+        # we limit filename to 120 characters
         # remove '-' from first_part because it is handled by cropName that manages max_length
         # and it behaves weirdly if it encounters '-'
         # moreover avoid more than one blank space at a time
@@ -270,6 +279,10 @@ class DocumentGenerationView(BrowserView):
         if output_format == "csv":
             csvOptions = CsvOptions(fieldSeparator=pod_template.csv_field_delimiter,
                                     textDelimiter=pod_template.csv_string_delimiter)
+        evaluator = None
+        if os.getenv('DOCUMENTGENERATOR_USE_COMPROMISER', True):
+            evaluator = Compromiser()
+
         renderer = Renderer(
             StringIO(document_template.data),
             generation_context,
@@ -283,6 +296,7 @@ class DocumentGenerationView(BrowserView):
             html=True,
             optimalColumnWidths=optimalColumnWidths,
             distributeColumns=distributeColumns,
+            evaluator=evaluator,
             stylesMapping=stylesMapping,
             stream=config.get_use_stream(),
             csvOptions=csvOptions,